==> 01.2-OptTableCombine-020999.ipynb <==
Copyright 2021 D. Bazzett  
The purpose of this code is to look at the COMIDs associated with the HWM data and create a custom hydrogeo-table (rating curve) for only these COMIDs. 99% of the data in the default tables is not included in the HWM data, so creating a custom table speeds up calculation times.

import pandas as pd
import numpy as np

# for comid in ['020301', '020401', '020402', '020403']:
comid = '020301'
hydrotable_file = "H:/My Drive/" + comid + "/hydrogeo-fulltable-" + comid + ".csv"
df1 = pd.read_csv(hydrotable_file)
print(len(df1.index))

for comid in ['020401', '020402', '020403']:
    hydrotable_file = "H:/My Drive/" + comid + "/hydrogeo-fulltable-" + comid + ".csv"
    df2 = pd.read_csv(hydrotable_file)
    df1 = df1.append(df2)
    print(len(df2.index))
df1

# hwms = pd.read_csv('./Cleaned Hwm/HWM_Combined_NoNegatives_Cleaned.csv')
# hwms = pd.read_csv('H:/My Drive/CODE/HAND/Cleaned Hwm/RAW/USGS_Somerset_DEM_HAND_HUC_calc_no-neg_clean.csv')
# hwms = pd.read_csv('H:/My Drive/CODE/HAND/HWM/USGS_Somerset_opt.csv')
hwms = pd.read_csv('H:/My Drive/CODE/HAND/HWM/USGS_Somerset_opt.csv')
hwms

# for comid in hwms.COMID.unique():
#     subset = hwms[hwms.COMID == comid]
#     print(subset.group_avg.mean())
#     if not np.isnan(subset.group_avg.mean()):
#         print('completed')
#         continue
#     print(comid)

# maxflow = pd.read_csv('H:/My Drive/CODE/HAND/NWM Flow/maxflow_timefile.csv')
# hwms = hwms.merge(maxflow, on='COMID', how='left')
# hwms.to_csv('H:/My Drive/CODE/HAND/Cleaned Hwm/RAW/USGS_Somerset_opt.csv')

len(hwms.COMID.unique())

# df1.loc[df1.CatchId == 6251308, 'SLOPE'] = 0.05
# df1[df1.CatchId == 6251308]

# 020301
df1.loc[df1.CatchId == 6251308, 'SLOPE'] = 0.15 / 5223 # PASS1
df1.loc[df1.CatchId == 6251234, 'SLOPE'] = 0.36 / 6934 # PASS2 
df1.loc[df1.CatchId == 6251128, 'SLOPE'] = 1.27 / 1729 # PASS3
df1.loc[df1.CatchId == 6249784, 'SLOPE'] = 0.00005 # PASS7 - arbitrary because slope is almost zero
df1.loc[df1.CatchId == 6251122, 'SLOPE'] = 0.37 / (694 + 701) # PASS10
df1.loc[df1.CatchId == 6251144, 'SLOPE'] = 0.37 / (694 + 701) # PASS10
df1.loc[df1.CatchId == 6250358, 'SLOPE'] = 0.16 / 2744 # PASS16
df1.loc[df1.CatchId == 6250388, 'SLOPE'] = 0.71 / 3277 # PASS17
df1.loc[df1.CatchId == 6250418, 'SLOPE'] = 1.40 / 2824 # PASS18

df1.loc[df1.CatchId == 6246206, 'SLOPE'] = 1.59 / 2824 # WANQ2
df1.loc[df1.CatchId == 6250622, 'SLOPE'] = 1.41 / 2480 # POMP4/5
df1.loc[df1.CatchId == 6249738, 'SLOPE'] = 0.00005     # POMP2 - arbitrary value to remove error
# df1[df1.CatchId == 6246098, 'SLOPE'] = 0.88 / 1013 # RAMA5 - 
df1.loc[df1.CatchId == 932030072, 'SLOPE']=1.75 / 635  # RAMA5
df1.loc[df1.CatchId == 6250680, 'SLOPE'] = 0.00005     # ROCK2 - arbitrary value to remove error

# 020402
df1.loc[df1.CatchId == 4491888, 'SLOPE'] = 0.5 / 3571 # NBRAN4

df1.loc[df1.SLOPE < 0.00003, 'SLOPE'] = 0.00003

df1['Discharge (m3s-1)'] = df1['WetArea (m2)'] * pow(df1['HydraulicRadius (m)'],2.0/3) * \
    pow(df1['SLOPE'],0.5) / df1['Roughness']

# df1[df1.CatchId == 4491888]

hwms = hwms.dropna(subset=['COMID'])
valid_rows1 = df1[df1['CatchId'].isin(set(hwms.COMID.unique()))]  
valid_rows1

# valid_rows1.to_csv('H:/My Drive/CODE/HAND/HWM/hydrogeo-fulltable-020999.csv')

gages = pd.read_csv('H:/My Drive/CODE/HAND/HWM/GagesNJ_Height_Flow_Cleaned.csv')
gages

valid_rows2 = df1[df1['CatchId'].isin(set(gages.COMID.unique()))]  
valid_rows2

valid_rows2.to_csv('H:/My Drive/CODE/HAND/HWM/hydrogeo-fulltable-020998.csv')
