==> 01.3-COMID-Regression-Data.ipynb <==
The purpose of this code is to aggregate all the data from every catchment in the 5x HUC6 units and keep just the first row of data for each catchment. Additional data created in ArcGIS for Land Use,  Runoff, and Centroid Coordinates is added in.

import pandas as pd

comid = '020200'
hydrotable_file = "H:/My Drive/" + comid + "/hydrogeo-fulltable-" + comid + ".csv"
df1 = pd.read_csv(hydrotable_file)
print(f'{comid} has {len(df1.index)} rows')

for comid in ['020301', '020401', '020402', '020403']:
    hydrotable_file = "H:/My Drive/" + comid + "/hydrogeo-fulltable-" + comid + ".csv"
    df2 = pd.read_csv(hydrotable_file)
    df1 = df1.append(df2)
    print(f'{comid} has {len(df2.index)} rows')
df1

df1 = df1.drop_duplicates(subset='CatchId')
df1

order = pd.read_csv('H:/My Drive/HAND/NHDPlusAttributes/PlusFlowlineVAA.csv')
df1 = df1.merge(order[['ComID','StreamOrde']], left_on='CatchId', right_on='ComID', how='left')
df1

lc = pd.read_csv('H:/My Drive/CODE/HAND/Catchment Data/LatLon-LC-RC.csv')
lc

df1 = df1.merge(lc[['gridcode','POINT_X','POINT_Y','AvgElev','RunoffC','MajLC']], left_on='CatchId', right_on='gridcode', how='left')
df1

df3 = df1.drop(labels=['Stage', 'Number of Cells', 'SurfaceArea (m2)', 'BedArea (m2)', 'Volume (m3)', \
                       'Roughness', 'TopWidth (m)', 'WettedPerimeter (m)', 'WetArea (m2)', \
                       'HydraulicRadius (m)', 'Discharge (m3s-1)','ComID','gridcode'], axis='columns')

df3 = df3.replace(to_replace='<Null>',value=float('NaN'))
df3

df3

df3.to_csv('H:/My Drive/CODE/HAND/Catchment Data/Catchment_Regression_Data.csv',index=False)

df3.describe()

df3.dropna()

nj = pd.read_excel('./Catchment Data/NJ_Catchments.xlsx')
nj

all_data = pd.read_csv('./Catchment Data/Catchment_Regression_Data.csv')
all_data

combo = nj[['gridcode']].merge(all_data,how='left',left_on='gridcode',right_on='CatchId')
combo

combo.to_csv('./Catchment Data/Catchment_Regression_Data_NJ.csv',index=False)
