==> 02.1-NWM-CurrentData.ipynb <==
Copyright 2021 D Bazzett  
The purpose of this code is to examine the NWM data in the native .comp file (which is a NetCDF file). No changes are made here



# !pip install fiona
# !pip install netcdf4
# !pip install xarray

import xarray as xr
import pandas as pd
# import fiona
# import geopandas as gpd

datax = xr.open_dataset("./NC/201108281500.CHRTOUT_DOMAIN1.comp")
datax

df = datax.to_dataframe()
df = df.reset_index()
df.loc[df.feature_id == 6249912]

df

datax['streamflow'].attrs

datax.streamflow

data = xr.open_dataset("nwm.t00z.short_range.channel_rt.f001.conus.nc")
data

for var in data.variables.values():
    print(var)

data.variables

df = data.to_dataframe()
df = df.reset_index(drop=False)
df

df.feature_id

comid = pd.read_csv("020301_unique.csv", header=None)
comid = comid.sort_values(by=0)
comid = comid.reset_index(drop=True)
comid = comid.rename(columns={0:'id'})
comid

df = df[df['feature_id'].isin(comid.id)]
df

# %%cmd
# python C:/Users/David/G-Rutgers/GitHub/nfie-floodmap/test/catchShapeByHUC.py 020301 C:/Users/David/G-Rutgers/WHIRL/HAND/020301/020301-flows.shp 020301-flows 9442367 C:/Users/David/G-Rutgers/WHIRL/HAND/NHDPlusV21_National_Seamless_Flattened_Lower48.gdb Catchment C:/Users/David/G-Rutgers/WHIRL/HAND/test 

%%cmd
python C:/Users/David/G-Rutgers/GitHub/nfie-floodmap/test/forecast-table.py C:\Users\David\G-Rutgers\WHIRL\HAND\020301\hydrogeo-fulltable-020301.csv C:\Users\David\G-Rutgers\GitHub\HAND\nwm.t00z.short_range.channel_rt.f001.conus.nc C:\Users\David\G-Rutgers\WHIRL\HAND\test C:\Users\David\G-Rutgers\WHIRL\HAND\NHDPlusV21_MA_02_NHDSnapshotFGDB_04\NHDSnapshot.gdb
