==> 02.2-NWM-ReanData.ipynb <==
Copyright 2021 D Bazzett  
The purpose of this code is to read and analyze Nat'l Water Model (NWM) Reanalyss Data.

Get reanalysis data from NWM through this link (hosted through AWS)
https://docs.opendata.aws/nwm-archive/readme.html


import pandas as pd
import datetime as dt
import h5py    
import numpy as np    
import xarray as xr

# nwm = pd.read_hdf('./200504040000.CHRTOUT_DOMAIN1.comp')
# This doesn't work - I have read that only very specifc HDF5 formats can be translated directly to pandas
# Here's an easy fix from stachexchange
# df = pd.DataFrame(np.array(h5py.File(path)['variable_1']))

f1 = h5py.File('./201108281500.CHRTOUT_DOMAIN1.comp','r+')    

f1.attrs.keys()

f1.attrs['stream_order_output']

f1['streamflow'].attrs

# f1.attrs['streamflow']

f1.keys

f1.keys()

# f1['reference_time'].dtype

f1['streamflow']

f1['streamflow'].attrs

f1.attrs

f1.attrs.keys()['_NCproperties']

f1.attrs['_NCProperties']

for item in f1.attrs.keys():
    print(item + ":", f1.attrs[item])
