==> 03.2-MaxflowToNC.ipynb <==
The purpose of this code is a workspace to convert the csv of maxflow into a netcdf file that can be used as input into the forecast-map script. 

import xarray as xr
import pandas as pd

datax = xr.open_dataset('./NC/inun-hq-table-at-20110701_000000-for-20110827_000000.nc')
datax

datax.attrs

for var in datax.variables.values():
    print(var)

file = './Smoothing/maxflow_default.csv'
flow = pd.read_csv(file)
flow

flowx = flow.to_xarray()
flowx.attrs['Timestamp'] = '20110701_000000'
flowx

for var in flowx.variables.values():
    print(var)

file_nc = file[:-4] + '.nc'
flowx.to_netcdf(file_nc)

new_file = xr.open_dataset(file_nc)
new_file
