==> 05-Process-HWM.ipynb <==
# !python -V
# !pip install dbfread

import osgeo
import shapefile as shp
import pandas as pd
from simpledbf import Dbf5
import numpy as np
import dbfread as dbfr

# Merge peak times
# data = pd.read_csv('./020301_020402_HWM.csv')
# dates = pd.read_csv('./HWM_PeakTimes.csv')
# dates.time_fmax = pd.to_datetime(dates.time_fmax, format='%Y%m%d_%H_Q_x',)
# data = data.merge(dates, on='COMID')
# data.to_csv('./020301_020402_HWM.csv')
# data

# dbf_file = 'C:/Users/David/G-Rutgers/HAND/ValidationDocs/USGS_HMW_Aug2011_Irene/USGS_HMW_Aug2011_Irene.dbf'
# In ArcGIS, take Irene HWM shapefile, reproject shapefile to NFIE DEM, Extract Values to Points
# (DEM to HWM, HAND to HWM, COMID to HWM) 
# Export new layer to dbf

# dbf_file = './NJ_HWM.dbf'
# dbf_file = 'C:/Users/David/G-Rutgers/HAND/ValidationDocs/Irene/HWM_DEM_HAND_COMID_020301.dbf'
# dbf = Dbf5(dbf_file)
# df1 = dbf.to_dataframe()

# df1 = df1.rename(columns={'ELEV_NAVD8':'HWM_NAVD8_ft'})
# df1 = df1.rename(columns={'DEM':'DEM_m'})
# df1 = df1.rename(columns={'HAND':'HAND_m'})

# # Convert HWM elev to m
# df1['HWM_NAVD8_m'] = ''
# df1['HWM_NAVD8_m'] = df1['HWM_NAVD8_ft'] * 0.3048

# # Calculate water depth at the location of HWM
# df1['Inun_Depth_m'] = ''
# df1['Inun_Depth_m'] = df1['HWM_NAVD8_m'] - df1['DEM_m']

# # Drop missing data - see the second dataset
# df2 = df1[df1.HAND_m > 0]
# df2 = df2[df2.HWM_NAVD8_ft > 0]
# # df2 = df2.reset_index(drop=True)

# # Calculate the total water depth based on HAND at HWM and Inun_Depth at 
# df2['Stream_H_m'] = ''
# df2['Stream_H_m'] = df2.Inun_Depth_m + df2.HAND_m

# # Take a look at the points with negative Inun_Depth values - any 
# # df_neg = df[df.Inun_Depth_m < 0]
# # pd.options.display.max_columns = None
# # pd.options.display.max_rows = None
# # df_neg

# # df2.to_csv('./NJ_HWM_USGS.csv')
# df2.head()

# import matplotlib.pyplot as plt
# plt.hist(df.Inun_Depth_m, bins = 10)
# plt.show()

# 958 HWM points total
# 932 HWM points where HWM_NAVD8_ft > 0
# 790 HWM points combined in 020301 and 020402 where HWM_NAVD8_ft > 0 and HAND > 0

# hwm = pd.read_csv('./HWM_combined.csv')
# hwm

# nhd_file = 'C:/Users/David/G-Rutgers/NHDPlusAttributes/PlusFlowlineVAA.dbf'
# nhd = pd.DataFrame(iter(dbfr.read(nhd_file)))
# nhd = nhd.rename(columns={'ComID':'COMID'})

# # Intermediate slice
# dfy = nhd[['StreamOrde','COMID']].copy()

# hwm = hwm.merge(dfy, on='COMID') #left_on='COMID', right_on='COMID'
# hwm.to_csv('./HWM_combined.csv')
# hwm

hwm.COMID.unique().size

df_a = pd.read_csv('C:/Users/David/G-Rutgers/020301/hydrogeo-fulltable-020301-updated.csv')
df_b = pd.read_csv('C:/Users/David/G-Rutgers/020401/hydrogeo-fulltable-020401-updated.csv')
df_c = pd.read_csv('C:/Users/David/G-Rutgers/020402/hydrogeo-fulltable-020402-updated.csv')
df_d = pd.read_csv('C:/Users/David/G-Rutgers/020403/hydrogeo-fulltable-020403-updated.csv')


df_a = df_a.append(df_b)
df_a = df_a.append(df_c)
df_a = df_a.append(df_d)
df_a

df_a = df_a.drop_duplicates(subset='CatchId', keep='first')
df_a = df_a.drop(labels=['Number of Cells', 'SurfaceArea (m2)','BedArea (m2)','Volume (m3)',\
                        'Roughness', 'TopWidth (m)', 'WettedPerimeter (m)','WetArea (m2)', 'HydraulicRadius (m)',\
                         'Discharge (m3s-1)', 'StreamOrde', 'rough'], axis=1)
df_a

extra = pd.read_csv('Catchments_LandUse.csv')
extra

df_a = df_a.rename(columns={'CatchId':'COMID'})
combine = df_a.merge(extra, on='COMID')
combine

combine = combine.drop_duplicates(subset='COMID', keep='first')
combine

combine.to_csv('./Catchment_Data.csv')

comids = df_a.CatchId.unique()
comids.size
# comids

comids.sort()
comids

# sorted(comids, key=lambda x: int(x[-1]))

comids.dtype

for item in comids:
    if isinstance(item, int):
        print(item)

valid_rows = df_a[df_a['CatchId'].isin(set(hwm.COMID.unique()))]  
valid_rows = valid_rows.drop(['Unnamed: 0'],axis=1)
# valid_rows.to_csv('./validation_hydrotable.csv')
valid_rows

valid_rows.columns

# # rasterio installed from Gohlke's binaries 
# import rasterio
# import numpy as np
# from affine import Affine
# from pyproj import Proj, transform

# import rasterio.features
# import rasterio.warp
# from matplotlib import pyplot
# from rasterio.plot import show
# import pandas as pd
# import numpy as np

# fname = 'C:/Users/David/G-Rutgers/020301/020301.tif'

# img = rasterio.open(fname)

# print(img.count)
# print("W = " + str(img.width) + ", H = " + str(img.height))
# print(img.bounds)
# band1 = img.read(1)
# print(band1[17289,19023])

# lat = pd.DataFrame()
# lon = 

# show(band1,0)

# pd.DataFrame(array.reshape([3,-1]).T)

# # Read raster
# with rasterio.open(fname) as r:
#     T0 = r.transform  # upper-left pixel corner affine transform
#     p1 = Proj(r.crs)
#     A = r.read()  # pixel values

# # All rows and columns
# cols, rows = np.meshgrid(np.arange(A.shape[2]), np.arange(A.shape[1]))

# # Get affine transform for pixel centres
# T1 = T0 * Affine.translation(0.5, 0.5)
# # Function to convert pixel row/column index (from 0) to easting/northing at centre
# rc2en = lambda r, c: (c, r) * T1

# # All eastings and northings (there is probably a faster way to do this)
# eastings, northings = np.vectorize(rc2en, otypes=[np.float, np.float])(rows, cols)

# # Project all longitudes, latitudes
# p2 = Proj(proj='latlong',datum='WGS84')
# longs, lats = transform(p1, p2, eastings, northings)
