==> 05.1-USGSHeightFormatting.ipynb <==
import pandas as pd
usgs = pd.read_csv('./USGS Flow/GagesNJ_Height.csv')
usgs

flow = pd.read_excel('./NWM Flow/all_flows.xlsx')
flow

when = flow[['COMID', 'peak_time']]
when

both = usgs.merge(when,on='COMID')
both

both.to_csv('./USGS Flow/GagesNJ_Height_Flow.csv')
