==> 06-ManningCalibration.ipynb <==
Copyright D Bazzett 2021  
The purpose of this code is to compare the formatted HWM data to the Rating Curve tables and calibrate the Manning's N for optimization

import pandas as pd

fc_path = './manning/'

# Read HWM and maxflow data
hwm = pd.read_csv('./020301_020402_HWM.csv')
forecast = pd.read_csv(fc_path + 'maxflow.csv')

# Join
join = hwm.merge(forecast, on='COMID')
join = join.rename(columns={'H':'Predict_Stream_H_m'})
join['rmse'] = ''

rough = pd.read_csv('roughness.csv')
rough['rmse'] = ''

for val in join.StreamOrde.unique():
    sample=join[join.StreamOrde == val]
    rmse = ((sample.Stream_H_m - sample.Predict_Stream_H_m) ** 2).mean() ** .5
    join.rmse[join.StreamOrde == val] = rmse
    rough.rmse[rough.StreamOrde == val] = rmse
    print("Order : " + str(val) + ", RMSE : " + str(rmse))
    print("  Avg H Observed:" + str(sample.Stream_H_m.mean()))
    print("  Avg H Predict: " + str(sample.Predict_Stream_H_m.mean()))

join

rough

list1 = rough.rough.unique()
list1.sort()
list1

import numpy as np
from scipy.optimize import minimize
def obj(x):
#     update_the_table(x)
    y=join.Stream_H_m
    y_hat=join.Predict_Stream_H_m
    rmse = ((join.Stream_H_m - join.Predict_Stream_H_m) ** 2).mean() ** .5
    return rmse

x0 = np.array(rough.rough)    #initial roughness, can be read from the original table
res = minimize(obj, x0, method='nelder-mead', options={'xatol': 1e-8, 'disp': True})
