==> 07-UpdateManning.ipynb <==
#!/usr/bin/env python
# coding: utf-8

import pandas as pd
import sys

def UpdateManning():
    # # Path to save to
    # path = './fc-nwm/Irene/manning02/'
    
    # Location of dummy file to update
    rough = pd.read_csv('./roughness.csv')
    
    rough.rough = float(manning)
    # rough.rough[rough.StreamOrde == 1] = 0.005
    # rough.rough[rough.StreamOrde == 2] = 0.005
    # rough.rough[rough.StreamOrde == 3] = 0.005
    # rough.rough[rough.StreamOrde == 4] = 0.005
    # rough.rough[rough.StreamOrde == 5] = 0.005
    # rough.rough[rough.StreamOrde == 6] = 0.005
    # rough.rough[rough.StreamOrde == 7] = 0.005
    
    rough.to_csv(path + '/roughness.csv')
    print("Roughness table created with roughness = " + manning)
    
    # Location of updated.csv to copy 
    # ht_file = "./fc-nwm/Irene/validation_hydrotable.csv"
    ht_file1 = "./020402/hydrogeo-fulltable-020402.csv"
    ht_file2 = "./020301/hydrogeo-fulltable-020301.csv"
    
    files = {ht_file1, ht_file2}
    
    for ht_file in files:
        print("Reading file " + ht_file)
        ht = pd.read_csv(ht_file)
        ht = ht.drop(['rough'], axis=1)
        
        
        df3 = ht.merge(rough, on='StreamOrde')
        
        # Calculate flow based on the new Manning's N
        df3['Discharge (m3s-1)'] = df3['WetArea (m2)'] * pow(df3['HydraulicRadius (m)'],2.0/3) * \
            pow(df3['SLOPE'],0.5) / df3['rough']
        
        # TODO: Do we overwrite the existing hydrotable or create a new one?
        # Write new table to file (comment out for testing)
        new_file = path + ht_file[1:] # [9:] 
        df3.to_csv(new_file)
        print('UpdateManning complete for ' + new_file)
    
    
if __name__ == '__main__':
    path    = sys.argv[1] # saveto path
    # ht_file = sys.argv[2] # rating curve table
    manning = sys.argv[2] # roughness
    UpdateManning()
