==> 08-SingleOptimization.ipynb <==
Copyright 2021 David Bazzett  
The purpose of this script is to optimize the Manning's roughness for an individual catchment. The rating curve must be calculated, the forecast must be generated, and the forecast height must be compared to the observed height. This must be iterated until an optimal Manning's value is reached.

#!/usr/bin/env python
# coding: utf-8

import pandas as pd
import sys

def UpdateManning():
    
    # Read dummy file, update roughness, write new file
#     rough = pd.read_csv('./roughness.csv')
#     rough.rough = float(manning)
#     rough.to_csv(path + '/roughness.csv')
#     print("Roughness table created with roughness = " + manning)
    
    # Location of updated.csv to copy 
    ht_file1 = "./020402/hydrogeo-fulltable-020402.csv"
    ht_file2 = "./020301/hydrogeo-fulltable-020301.csv"
    
    files = {ht_file1, ht_file2}
    
    for ht_file in files:
        print("Reading file " + ht_file)
        ht = pd.read_csv(ht_file)
#         print(ht.head())
        if opt_comid not in ht.CatchId.unique():
            break
        else:
            print("COMID " + str(opt_comid) + " found in " + ht_file)
            
        subset = ht.Roughness.loc[ht.CatchId == opt_comid]
                
        # Calculate flow based on the new Manning's N
        ht['Discharge (m3s-1)'] = ht['WetArea (m2)'] * pow(ht['HydraulicRadius (m)'],2.0/3) * \
            pow(ht['SLOPE'],0.5) / ht['Roughness']
        
        # Write new table to file (comment out for testing)
#         new_file = path + ht_file[1:] # [9:] 
        ht.to_csv(ht_file)
#         print('UpdateManning complete for ' + new_file)

        # This was test code to see if it was faster to calculate for a smaller sample instead of the whole table.
#         subset = ht.loc[ht.CatchId == opt_comid]
#         subset.Roughness = rough  
#         # Calculate flow based on the new Manning's N
#         subset['Discharge (m3s-1)'] = subset['WetArea (m2)'] * pow(ht['HydraulicRadius (m)'],2.0/3) * \
#             pow(ht['SLOPE'],0.5) / subset['Roughness']
#         ht['Discharge (m3s-1)'].loc[ht.CatchId == opt_comid] = subset['Discharge (m3s-1)']
        

hwm = pd.read_csv('./020301_020402_HWM.csv')
hwm = hwm.head()
hwm_comid = hwm.COMID.unique()
rough=0.2
for opt_comid in hwm_comid:
    print(opt_comid)
#     UpdateManning()
    print(hwm.time_file.loc[hwm.COMID == opt_comid].unique())

import numpy as np
from scipy.optimize import minimize
def obj(x):
#     update_the_table(x)
    y=join.Stream_H_m
    y_hat=join.Predict_Stream_H_m
    rmse = ((join.Stream_H_m - join.Predict_Stream_H_m) ** 2).mean() ** .5
    return rmse

x0 = np.array(rough.rough)    #initial roughness, can be read from the original table
res = minimize(obj, x0, method='nelder-mead', options={'xatol': 1e-8, 'disp': True})

if __name__ == '__main__':
    path    = sys.argv[1] # saveto path
    # ht_file = sys.argv[2] # rating curve table
    manning = sys.argv[2] # roughness
    UpdateManning()
