==> 20.1-HeadOfTide.ipynb <==
Copyright 2021 D Bazzett  
The purpose of this code is to determine which catchments have tidal influence as determined by the NJ Head of Tide (HOT) Survey. HOT points have catchment data added in ArcGIS. We then create a map of downstream catchments using the NHDPlus Value Added Attribute (VAA) dataset.

import dbfread as dbf
import numpy as np
import pandas as pd
pd.set_option('display.max_columns',15)

# These are the attributes for the HUC02 of 02, containing 0201*, 0202*, 0203*, etc. 
# All of NJ is contained in this file
nhd_file = 'H:/My Drive/HAND/NHDPlusAttributes/PlusFlowlineVAA.dbf'
nhd = pd.DataFrame(iter(dbf.read(nhd_file)))

# csv_name = nhd_file[:-4]+'.csv'
# nhd.to_csv(csv_name,index=False)

# nhd = pd.read_csv('H:/My Drive/HAND/NHDPlusAttributes/PlusFlow.csv')
nhd

tidal = nhd[['ComID','Tidal']]
tidal.to_csv('H:/My Drive/HAND/NHDPlusAttributes/Tidal.csv',index=False)
file = 'H:/My Drive/ArcGIS-NJ-HAND/Tide/020301.csv'
reclass = pd.read_csv(file)
reclass = reclass.merge(tidal, left_on='FROM_', right_on='ComID', how='left')
reclass.to_csv(file[:-4] + 'tide-vaa.csv', index=False)

hot = pd.read_csv('H:/My Drive/ArcGIS-NJ-HAND/Tide/Head_of_Tide_DD_catch.csv')
hot

print(f'{len(hot.CatchId.unique())} unique HoT catchments')

test = hot.CatchId.unique()[0]
test

nhd.loc[nhd.FROMCOMID == 9481590, 'TOCOMID'].values[0]

nhd.loc[nhd.FROMCOMID == 9480456, 'TOCOMID'].values[0]

lglist = []
if False:
    for pt in hot.CatchId.unique():
        print(f'{pt} has ',end='')
        step1 = pt
        smlist = [step1]
        while len(nhd.loc[nhd.FROMCOMID == step1, 'TOCOMID'].values) != 0:
            step2 = nhd.loc[nhd.FROMCOMID == step1, 'TOCOMID'].values[0]
    #         print(step2)
            smlist.append(step2)
            step1 = step2
        print(f'{len(smlist)} steps, ',end='')
        lglist.extend(smlist)
        print(f'{len(np.unique(lglist))} total catchments so far...')

tidals = np.unique(lglist)
tidals

print(tidals)

# np.savetxt('./HeadOfTide/tidals.csv', tidals, delimiter=',')

tidals = np.genfromtxt('./HeadOfTide/tidals.csv', delimiter=',')
tidals

reclass = pd.read_csv('H:/My Drive/ArcGIS-NJ-HAND/Tide/020301.csv')
reclass

reclass['OUT'] = 0
reclass

reclass.loc[reclass.FROM_.isin(tidals),'OUT'] = 1
reclass.loc[reclass.FROM_.isin(tidals)]

reclass.to_csv('H:/My Drive/ArcGIS-NJ-HAND/Tide/020301tide.csv',index=False)
