==> 30.1-ModelFlowToNC.ipynb <==
Copyright 2021 D Bazzett  
The purpose of this code is to convert our flow forecasts for 10-year and worst case + 10%, 20%, and 30% into friendly netcdf files that can be read by Yan's code.

# !pip install pandas==0.25

import pandas as pd
print(pd.__version__)
import xarray as xr


# file = './ModelFlow/MODEL_FLOW.csv'
file = './ModelFlow/res_linear.csv'
flow = pd.read_csv(file,encoding = 'unicode_escape')
flow

len(flow.index)

from datetime import datetime

names = ['dis_a00', 'dis_a10', 'dis_a20', 'dis_a30', 'dis_a40', 'dis_a50',\
         'dis_w00', 'dis_w10', 'dis_w20', 'dis_w30', 'dis_w40', 'dis_w50',]

x=0
y=1
for name in names:
    set_time = datetime(2020, 1, 1, 1, 0, 0)
    print(name,end='')
    subset = flow[['COMID',name]]
    subset = subset.rename(columns={'COMID': 'feature_id', name: 'streamflow'})
    
    flowx = xr.Dataset(
            data_vars=dict(
                streamflow = (['feature_id'], subset.streamflow.values)
            ),
            coords=dict(
                time=(['time'],[set_time]),
                reference_time=(['reference_time'],[set_time]),
                feature_id=subset.feature_id.values,
            ),
            attrs=dict(
            model_output_valid_time = '2020-01-0' + str(y) + '_0' + str(x) + ':00:00',
            model_initialization_time = '2020-01-01_00:00:00',
            )
    ) 
    print(' 2020-01-0' + str(y) + '_0' + str(x) + ':00:00')
    flowx.to_netcdf('./ModelFlow/' + name + '.nc')
    x += 1
    if x == 6:
        y = 2
        x = 0

# Dimensions     feature_id: 2729077 reference_time: 1 time: 1
flowx

for var in flowx.variables.values():
    print(var)
