==> 40-AvgHeight.ipynb <==
# !pip install netcdf4

import pandas as pd
import numpy as np
import xarray as xr

# path = 'H:/My Drive/ArcGIS-NJ-HAND/Strahler'

# huc6 = '020200'
# file = path + '/' + huc6 + 'flow.csv'
# tables = pd.read_csv(file)


# for huc6 in ['020200', '020301', '020401', '020402', '020403']:
#     file = path + '/' + huc6 + 'flow.csv'
#     table = pd.read_csv(file)
#     tables = tables.append(table)

# tables

# tables.to_csv('./MapAvg/All_flowlines.csv',index=False)

# subset = tables[['COMID','StreamOrde','UpHydroseq','Hydroseq','DnHydroseq']]
# subset = tables[['COMID','StreamOrde','FromNode','ToNode']]
# subset.to_csv('./MapAvg/All_Sequence.csv',index=False)
# subset.head()

table = pd.read_csv('./MapAvg/All_Sequence.csv')
# table = subset.copy()
table.head()

xfile = './MapAvg/inun-hq-table-at-20200101_000000-for-20200101_000000.nc'
# xfile = './MapAvg/inun-hq-table-at-20200101_000000-for-20200101_000000_avg.nc'

datax = xr.open_dataset(xfile)
xr_att = datax.attrs
datax

datapd = datax.to_dataframe()
# datapd.reset_index()
print(len(datapd))
datapd.head(20)

# data2 = datapd.merge(table[['COMID','StreamOrde','UpHydroseq','Hydroseq','DnHydroseq']], on='COMID', how='left')
# # data2['UpH']  = ''
# # data2['UpOrde'] = ''
# # data2 = data2.merge(table[['COMID','ToNode']], on='COMID', how='left')
# # data2['DnH']  = ''
# # data2['DnOrde'] = ''
# # data2

# data2 = data2.merge(data2[['UpHydroseq','H']], left_on='Hydroseq', right_on='UpHydroseq', how='left')
# data2 = data2.rename(columns={'UpHydroseq_x': 'UpHydroseq', 'H_x': 'H', 'H_y': 'UpH'})
# # data2 = data2.drop(['UpHydroseq_y'], axis='columns')
# data2 = data2.merge(data2[['DnHydroseq','H']], left_on='Hydroseq', right_on='DnHydroseq', how='left')
# data2 = data2.rename(columns={'DnHydroseq_x': 'DnHydroseq', 'H_x': 'H', 'H_y': 'DnH'})
# # data2 = data2.drop(['DnHydroseq_y'], axis='columns')
# data2

data2 = datapd.merge(table, on='COMID', how='left')

data2 = data2.merge(data2[['FromNode','H','StreamOrde','ToNode']], left_on='ToNode', right_on='FromNode', how='left')
data2 = data2.rename(columns={'FromNode_x': 'FromNode', 'ToNode_x': 'ToNode',\
                              'StreamOrde_x': 'StreamOrde', 'StreamOrde_y': 'ToStreamOrde', \
                              'H_x': 'H', 'H_y': 'ToH', 'ToNode_y':'ToToNode'})
data2 = data2.drop(['FromNode_y'], axis='columns')

data2 = data2.merge(data2[['FromNode','H','StreamOrde']], left_on='ToToNode', right_on='FromNode', how='left')
data2 = data2.rename(columns={'FromNode_x': 'FromNode', 'ToNode_x': 'ToNode',\
                              'StreamOrde_x': 'StreamOrde', 'StreamOrde_y': 'ToToStreamOrde', \
                              'H_x': 'H', 'H_y': 'ToToH',}) # 'ToNode_y':'ToToNode'})
data2 = data2.drop(['FromNode_y'], axis='columns')

data2 = data2.merge(data2[['ToNode','H','StreamOrde','FromNode']], left_on='FromNode', right_on='ToNode', how='left')
data2 = data2.rename(columns={'ToNode_x': 'ToNode', 'FromNode_x':'FromNode', \
                              'StreamOrde_x': 'StreamOrde', 'StreamOrde_y': 'FromStreamOrde', \
                              'H_x': 'H', 'H_y': 'FromH', 'FromNode_y':'FromFromNode'})
data2 = data2.drop(['ToNode_y'], axis='columns')

data2 = data2.merge(data2[['ToNode','H','StreamOrde']], left_on='FromFromNode', right_on='ToNode', how='left')
data2 = data2.rename(columns={'ToNode_x': 'ToNode', 'FromNode_x':'FromNode', \
                              'StreamOrde_x': 'StreamOrde', 'StreamOrde_y': 'FromFromStreamOrde', \
                              'H_x': 'H', 'H_y': 'FromFromH',}) # 'FromNode_y':'FromFromNode'})
data2 = data2.drop(['ToNode_y'], axis='columns')

print(len(data2))
data2.head(20)

test = data2.copy()
# test = test.dropna()
# test.dtypes
test['Avg3'] = ''
test['Avg5'] = ''
test.loc[test.FromStreamOrde == test.ToStreamOrde, 'Avg3' ] = (test.H + test.FromH + test.ToH) / 3
test.loc[test.FromFromStreamOrde == test.ToToStreamOrde, 'Avg5' ] = \
                    (test.FromFromH + test.FromH + test.H + test.ToH + test.ToToH) / 5
print(len(test))
test.head(20)

sub1 = test[['COMID','H','Avg3']]
sub1.H = sub1.Avg3
sub1 = sub1.drop_duplicates(subset='COMID')
sub1.head(20)

sub2 = test[['COMID','H','Avg5']]
sub2.H = sub2.Avg5
sub2 = sub2.drop_duplicates(subset='COMID')
sub2.head(20)

# test.loc[test.Avg3 != '', 'H'] = test.loc[test.Avg3 != '', 'Avg3']
# test.loc[test.Avg5 != '', 'H'] = test.loc[test.Avg5 != '', 'Avg5']
# test.head(20)

# test = test.sort_values('H', ascending=True).drop_duplicates('COMID').sort_index()
# test = test.reset_index(drop=True)
# test

test.loc[test.COMID == 6249686]

# chop = test[['COMID','H','Q']]
chop = datapd.copy()
chop = chop.merge(sub1, on='COMID')
chop = chop.merge(sub2, on='COMID')
print(len(chop))
chop.head(20)

# chop.loc[5] - chop.loc[6]

# chop = chop.drop_duplicates(subset='COMID')
# chop

chop.loc[chop.Avg3 != '', 'H_x'] = chop.loc[chop.Avg3 != '', 'Avg3']
chop.loc[chop.Avg5 != '', 'H_x'] = chop.loc[chop.Avg5 != '', 'Avg5']
chop

flowx = chop.to_xarray()
flowx.attrs = xr_att
# flowx = flowx.drop_vars('index')
flowx

# flowx.to_netcdf(xfile[:-3] + '_avg.nc')
